https://bitbucket.org/leonardofialho/opttran/issue/2/threshold
Adding a threshold parameter to macpo-analyze. The invocation
command line is now:

macpo-analyze [OPTIONS] [threshold] macpo-log-file

The threshold accepts a decimal value between 0 and 1, which is
interpreted as a percentage. Hence threshold=0.1 implies 10%.

The threshold percentage is applied to the number memory references.
Due to sampling, this value may deviate from the expected count
of references.

When left unspecified, the threshold is set to 0.1.