Added modified C++ style checker.
This patch adds two files to the scripts/ directory. The cpplint.py script is a
modified incarnation of the Google C++ style guide chcker while the
format-check.sh script is a wrapper around cpplint.py.

When format-check.sh is invoked with the argument as "mod", cpplint.py is
called on all files modified since the last commit. Another option is to call
the format-check.sh script with the argument as "staged", which invokes
cpplint.py on the files about to be committed. The workflow in this case is to
run "git add foo bar" and then run "scripts/format-check.sh staged" *before*
running "git commit".

When invoked without any arguments, format-check defaults to checking all
modified files.