Reset loop header components for 2+ assignments.
This patch resets the index (idxv_expr) and increment (incr_expr) expressions
if the for-loop body assigns a value to the index variable when the increment
expression has already been identified. This task is already performed for
while and do-while loops, hence those loops do not need any special action.