Allow multiple --macpo: actions in one invocation.
This patch converts the ACTION_ enums into bitset values that are combined into
a single int16_t type. This allows multiple actions to be specified on the
command line. As a result, as long the instrumentations are compatible with
each other, they can be clubbed together into the same code.