Added set_endOfConstruct() calls.
Honestly, I am tired of seeing the set_endOfConstruct() == NULL error pop up during instrumentation! The remedy is to add a set_endOfConstruct(Sg_File_Info* file_info) call after every creation of an 'Sg...' object. To me, it seems like this call can easily be included in the constructor of the object and it eludes me why the designers of Rose chose to not add this call. This patch religiously (possibly unnecessarily) adds the set_endOfConstruct() call after every 'Sg...' constructor.

EOR (end of rant).