Change location delimiter from '#' to ':'.
I know, you probably think that I am obsessed with details. Well, I will always
contend that it is easier to memorize and use systems if we stick to the
standard. So while your scripts may have been coded to use '#' as the location
delimiter (to separate the function/file name from the line number), I wish to
reduce the confusion / pain for subsequent users and hence this change.
Henceforth, the correct way to indicate a loop to instrument is: foo:35. The
parsing code works for C++ specifiers (like foo::bar:3 or ::bar:3) as well. If
you aren't including a line number, you should be able to use your scripts
without any change.