Combine reuse distance metrics and non-temporal analysis.
Non-temporal stores are to be recommended only when there is a high chance that
the data being written will not be fetched back into memory in the near future.
This patch merges the reuse distance metrics from the previous patch to
recommend whether the #pragma vector nontemporal(var) statement should be
prepended to the loop.

The reuse distance analysis was modified to collect information about
write-only arrays.