Fork a new process for each integration test.
Previously, each test would be executed in the same process as the test
harness, which was fine for unit tests. For integration tests, however, this
does not work because Rose seems to maintain some global state which isn't
reset on each invocation of new/delete SgProject(). This patch creates and
reaps a new process for each integration test. It isn't yet known if this will
fix all problems (because the tests that fail intermittently will be included
in a subsequent patch).