Check whether the alignment check instrumentation depends on the input values.
This patch gets the index expression of the array references that are instrumented for alignment checks and tests whether these depend on constants or non-constants. It uses reaching definition analysis to get the definition of the index variable and recurses until it reaches a constant, a variable of type constant, the function arguments or an unhandled type.