Print function names and line numbers in recommendations.
This patch passes function addresses to instrumentation functions and
reconstructs the function name at the time of the recommendations. This patch
is an extension of a previous patch [a7151fbc9dacfe8c1b2712a8b5a9115a87ee2a82 /
Allow 2+ actions and locations in instrumentation]. An additional patch will be
issued to print the filename and line number instead of the function name.