Print output in a format that is easy to parse.
This patch introduces a new environment variable ("MACPO_DISPLAY_BOT") that,
when set, causes the analysis results to be displayed in a less verbose (and
more ordered form). For instance:

$ MACPO_DISPLAY_BOT=true ./mm
0.0

[macpo]  compute.c:10|sstore_alignment|status = NOT_ALIGNED
[macpo]  compute.c:10|alignment|status = NOT_ALIGNED

This patch is an extension of the changes that were introduced in
[ f7a64b2ede83fcdf32c5aa565cc9f9a24908b27e / Translate function addresses to
filenames]. This change is the last of the three changes that were required for
enabling automation of vectorization-opportunities code.