Compute dominators for CFG.
This patch uses the CFG built by Rose to compute dominators (and immediate
dominators) for each node of the CFG. This information will be used to
determine which statements preceed which others during execution. The execution
sequence is important to understand which definitions are to be considered when
analyzing initial values of loop index variables.