Use DefUseTables instead of VariableRenaming.
This patch is part of a more elaborate change to move to using more precise
analysis. Until now, various parts of MACPO (specifically, those dealing with
loop index information analysis) used VariableRenaming objects to track reuse
distance information. However, the VariableRenaming information is built on top
of reaching definitions exported by DefUseTables and hence the results were not
as precise as we would have liked them to be. Furthermore, in what appears to
be a bug in Rose, the traditional MMM example crashed during compilation when
VariableRenaming was run on a project containing more than one source file.

The change to DefUseTables (specifically, just definitions and not uses) will
be used in combination with the immediate dominator analysis from an earlier
patch [a5f389601c5ba16e4a58d37186b846726e8a9a08 / Compute dominators for CFG]
to improve the precision of determining whether a future recommendation will be
affected by a different input than what was provided on the command line.