Ignore data points if histogram is not ready.
This patch removes the asserts that checked whether memory allocation was
successful before writing (or ignoring) the specific data point. Considering
that there are thousands (if not more) of measurements coming in, it is okay to
miss the initial few. Hence this patch skips recording the data point if the
histogram has not been fully allocated.