Allow specification of target compiler.
This patch adds a new option to macpo.sh [--macpo:compiler] which allows
specifying the compiler to be invoked to compile / link the binary. The
arguments passed to macpo.sh are munged to make them compatible with the target
compiler (specifically, -rose:openmp:ast_only is removed and file names are
changed to reflect the morphed i.e. unparsed filenames). This makes it possible
to use the Intel (or any other compiler) with MACPO instead of building a
separate version of Rose configured to use the Intel compiler as the target
compiler. Thanks to Madhu from UTEP for reminding the importance of this
change.